requirements.txt
astroid==2.2.5
autopep8==1.4.3
colorama==0.4.1
dj-database-url==0.5.0
Django==2.1.7
djangorestframework==3.9.2
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.7
pycodestyle==2.5.0
pylint==2.3.1
python-decouple==3.1
pytz==2018.9
six==1.12.0
typed-ast==1.3.1
wrapt==1.11.1
